Create the player class and your unit test.
This creates the Player class and its unit test.
The player class is in charge of creating a player
with his chip and two flags, the first is used to
know if the player is a winner and the second is
used to know if it is his turn to put his chip.