Create the Emulator class and its unit test.
This creates the Emulator class and its unit test.
The emulator class is in charge of simulating the
TicTacToe game, creating and configuring the game
environment.
This class controls the entry of the players position
on the board and checks if the position is empty
to add a tile.
Also, this checks when the game ends and sends a
message if there is a winner or a tie.